HIP 15 | Barney David | Refactor PatientExtensions and move to its own file
