+++ src/In.ProjectEKA.HipService/OpenMrs/OpenMrsPatientMatchingRepository.cs
@@ -6,6 +6,7 @@ namespace In.ProjectEKA.HipService.OpenMrs
     using HipLibrary.Matcher;
     using HipLibrary.Patient.Model;
     using DiscoveryRequest = HipLibrary.Patient.Model.DiscoveryRequest;
+    using In.ProjectEKA.HipService.OpenMrs.Mappings;
 
     public class OpenMrsPatientMatchingRepository : IMatchingRepository
     {
@@ -32,26 +33,6 @@ namespace In.ProjectEKA.HipService.OpenMrs
                     .Select(p => p.ToHipPatient(request.Patient?.Name))
                     .ToList()
                     .AsQueryable();
-        }
-    }
-
-    public static class PatientExtensions
-    {
-        public static Patient ToHipPatient(this Hl7.Fhir.Model.Patient openMrsPatient, string patientSearchedName)
-        {
-
-            return new Patient()
-            {
-                Name = patientSearchedName,
-                Gender = openMrsPatient.Gender.HasValue ? (Gender)((int)openMrsPatient.Gender) : (Gender?)null,
-                YearOfBirth = (ushort?)openMrsPatient.BirthDateElement?.ToDateTimeOffset()?.Year,
-                Identifier = openMrsPatient.Id
-           };
-        }
-
-        public static Hl7.Fhir.Model.AdministrativeGender? ToOpenMrsGender(this Gender? gender)
-        {
-            return gender.HasValue ? (Hl7.Fhir.Model.AdministrativeGender)((int)gender) : (Hl7.Fhir.Model.AdministrativeGender?)null;
         }
     }
 }
